Updates for Max 8.5.0
As of Max 8.5.0, the Node.js version was bumped from 16.6 to 20.6. This breaks the previous functionality of our index.js, which is now reconfigured to run as an ES module (index.mjs).  This also requires the node.script objects to specify the --experimental-modules flag. See socketio_demonstration.maxpat for details.